WIP: Schema definition is parsed
